+++ src/main/java/edu/ie3/datamodel/io/processor/Processor.java
@@ -43,7 +43,17 @@ public abstract class Processor<T> {
   /* Quantities associated to those fields must be treated differently (e.g. input and result), all other quantity /
    * field combinations can be treated on a common basis and therefore need no further distinction */
   private static final Set<String> specificQuantityFieldNames =
-      Set.of("eConsAnnual", "energy", "eStorage", "q", "p", "pMax", "pOwn", "pThermal");
+      Set.of(
+          "eConsAnnual",
+          "energy",
+          "eStorage",
+          "q",
+          "p",
+          "pMax",
+          "pOwn",
+          "pThermal",
+          "pRef",
+          "pMin");
 
   private static final GeoJsonWriter geoJsonWriter = new GeoJsonWriter();