+++ src/main/java/edu/ie3/datamodel/models/input/system/type/chargingpoint/ChargingPointTypeUtils.java
@@ -168,7 +168,8 @@ public class ChargingPointTypeUtils {
     ChargingPointType res;
 
     // valid regex for either custom or pre-defined types
-    String validCustomRegex = "(\\w+\\d*)\\s*\\(\\s*(\\d+\\.?\\d+)\\s*\\|\\s*(AC|DC)\\s*\\)";
+    String validCustomRegex =
+        "([\\w \\-]+)\\s*\\(\\s*(\\d+\\.?\\d+(?:E-?\\d+)?)\\s*\\|\\s*(AC|DC)\\s*\\)";
 
     // does it match the valid regex?
     Pattern pattern = Pattern.compile(validCustomRegex);